+++ cmn-data-svc-server/src/main/java/org/acumos/cds/service/SolutionSearchServiceImpl2.java
@@ -47,12 +47,30 @@ import org.springframework.stereotype.Service;
 
 /**
  * Hibernate-assisted methods to search solutions.
+ * <P>
+ * These two aspects must be observed to get pagination working as expected:
+ * <OL>
+ * <LI>For all to-many mappings, force use of separate select instead of left
+ * outer join. This is far less efficient due to repeated trips to the database.
+ * </LI>
+ * <LI>Specify an unambiguous ordering. This at least is cheap, just add
+ * order-by the ID field.</LI>
+ * </OL>
+ * I'm not the only one who has fought Hibernate to get paginated search
+ * results:
+ * 
+ * <PRE>
+ * https://stackoverflow.com/questions/300491/how-to-get-distinct-results-in-hibernate-with-joins-and-row-based-limiting-pagi
+ * https://stackoverflow.com/questions/9418268/hibernate-distinct-results-with-pagination
+ * https://stackoverflow.com/questions/11038234/pagination-with-hibernate-criteria-and-distinct-root-entity
+ * https://stackoverflow.com/questions/42910271/duplicate-records-with-hibernate-joins-and-pagination
+ * </PRE>
  */
-@Service("solutionSearchService")
+@Service("solutionSearchService2")
 @Transactional
-public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl implements SolutionSearchService {
+public class SolutionSearchServiceImpl2 extends AbstractSearchServiceImpl implements SolutionSearchService {
 
-	private final EELFLoggerDelegate logger = EELFLoggerDelegate.getLogger(SolutionSearchServiceImpl.class);
+	private final EELFLoggerDelegate logger = EELFLoggerDelegate.getLogger(SolutionSearchServiceImpl2.class);
 
 	@Autowired
 	private SessionFactory sessionFactory;
@@ -69,20 +87,24 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 		Criteria criteria = sessionFactory.getCurrentSession().createCriteria(MLPSolution.class);
 		super.buildCriteria(criteria, queryParameters, isOr);
 
+		// Adjust fetch mode on tags to block Hibernate from using left outer join,
+		// which builds a cross product that contains duplicate rows.
+		// This is a horrid violation of information hiding.
+		criteria.setFetchMode("tags", FetchMode.SELECT);
+
 		// Count the total rows
 		criteria.setProjection(Projections.rowCount());
-		// Request distinct; cross product yields multiple rows with same solution
-		criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY);
 		Long count = (Long) criteria.uniqueResult();
 		if (count == 0)
 			return new PageImpl<>(new ArrayList<>(), pageable, count);
 
 		// Reset the count criteria
 		criteria.setProjection(null);
-		// Still need distinct
-		criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY);
 		// Add pagination and sort
 		super.applyPageableCriteria(criteria, pageable);
+		// Fallback order on a unique field. Without this the pagination
+		// yields odd results; e.g., request 10 items but only get 8.
+		criteria.addOrder(Order.asc("solutionId"));
 
 		// Get a page of results and send it back with the total available
 		List<MLPSolution> items = criteria.list();
@@ -109,6 +131,13 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 		criteria.createAlias("owner", ownerAlias);
 		criteria.createAlias("tags", tagAlias);
 
+		// Adjust fetch mode to block Hibernate from using left outer join,
+		// which builds a cross product that contains duplicate rows.
+		// This is a horrid violation of information hiding.
+		criteria.setFetchMode("revisions", FetchMode.SELECT);
+		criteria.setFetchMode(revAlias + ".artifacts", FetchMode.SELECT);
+		criteria.setFetchMode("tags", FetchMode.SELECT);
+
 		// Attributes on the solution
 		criteria.add(Restrictions.eq("active", active));
 		if (nameKeywords != null && nameKeywords.length > 0)
@@ -127,20 +156,19 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 		if (tags != null && tags.length > 0)
 			criteria.add(Restrictions.in(tagAlias + ".tag", tags));
 
-		// Request count of rows
+		// Count the total rows. This does NOT run the left outer joins for tags etc.
 		criteria.setProjection(Projections.rowCount());
-		// Request distinct set, cross product yields repeats
-		criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY);
 		Long count = (Long) criteria.uniqueResult();
 		if (count == 0)
 			return new PageImpl<>(new ArrayList<>(), pageable, count);
 
-		// Remove the count projection
+		// Reset the count criteria
 		criteria.setProjection(null);
-		// Still need distinct
-		criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY);
 		// Add pagination and sort
 		super.applyPageableCriteria(criteria, pageable);
+		// Fallback order on a unique field. Without this the pagination
+		// yields odd results; e.g., request 10 items but only get 8.
+		criteria.addOrder(Order.asc("solutionId"));
 
 		// Get a page of results
 		List<MLPSolutionFOM> items = criteria.list();
@@ -154,7 +182,7 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 			if (item instanceof MLPSolutionFOM)
 				solutions.add(((MLPSolutionFOM) item).toMLPSolution());
 			else
-				throw new AssertionFailure("findPortalSolutions: unexpected type: {} " + item.getClass().getName());
+				throw new AssertionFailure("findPortalSolutions: unexpected type: " + item.getClass().getName());
 
 		logger.debug(EELFLoggerDelegate.debugLogger, "findPortalSolutions: result size={}", solutions.size());
 		return new PageImpl<>(solutions, pageable, count);
@@ -169,15 +197,18 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 			String[] validationStatusCode, Date date, Pageable pageable) {
 
 		Criteria criteria = sessionFactory.getCurrentSession().createCriteria(MLPSolutionFOM.class);
+
 		criteria.createAlias("revisions", revAlias);
 		criteria.createAlias(revAlias + ".artifacts", artAlias);
 
 		// Adjust fetch mode to block Hibernate from using left outer join,
 		// which builds a cross product that contains duplicate rows.
 		// This is a horrid violation of information hiding.
+		/*
 		criteria.setFetchMode("revisions", FetchMode.SELECT);
 		criteria.setFetchMode(revAlias + ".artifacts", FetchMode.SELECT);
-
+		*/
+		
 		criteria.add(Restrictions.eq("active", active));
 		if (accessTypeCode != null && accessTypeCode.length > 0)
 			criteria.add(Restrictions.in(revAlias + ".accessTypeCode", accessTypeCode));
@@ -194,7 +225,7 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 		itemModifiedAfter.add(revModified);
 		itemModifiedAfter.add(artModified);
 		criteria.add(itemModifiedAfter);
-
+		
 		// Request count of rows
 		criteria.setProjection(Projections.rowCount());
 		Long count = (Long) criteria.uniqueResult();
@@ -203,8 +234,6 @@ public class SolutionSearchServiceImpl extends AbstractSearchServiceImpl impleme
 
 		// Remove the count projection
 		criteria.setProjection(null);
-		// Without this the result becomes a list of (solution, revision, artifact)!
-		criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY);
 		// Add pagination and sort
 		super.applyPageableCriteria(criteria, pageable);
 		// Fallback order on a unique field. Without this the pagination